feat: Update routes and templates for dashboard changes
