+++ src/routes/auth.py
@@ -1,51 +1,23 @@
 import datetime
-from flask import Flask, render_template, redirect, url_for, request, blueprints, flash
-from flask_login import LoginManager, UserMixin, login_user, login_required, logout_user, current_user
-from flask_sqlalchemy import SQLAlchemy
+from flask import render_template, redirect, url_for, request, blueprints, flash
+from flask_login import LoginManager, login_user, login_required, logout_user, current_user
 from werkzeug.security import generate_password_hash, check_password_hash
 
-
-from flask import Flask, render_template, request, jsonify
-import subprocess
-
 from src.scripts.email_me import send_smpt_email
-
 from src.config import app, db
-from src.models import User, SmptEamilPasswordConfig, DashboardSettings
+from src.models import User, DashboardSettings
 from src.utils import render_template_from_file
+from src.routes.helper import get_email_addresses
 
 auth_bp = blueprints.Blueprint('auth', __name__)
 
-# Configure Flask-Login
 login_manager = LoginManager()
 login_manager.init_app(app)
 login_manager.login_view = 'login'
 
-# Define the User model
-
 @login_manager.user_loader
 def load_user(user_id):
     return User.query.get(int(user_id))
-
-
-def get_email_addresses(user_level=None, receive_email_alerts=True, fetch_all_users=False):
-    with app.app_context():
-        # Build query filter based on the presence of `user_level`
-        filters = []
-        if user_level:
-            filters.append(User.user_level == user_level)
-        if not fetch_all_users:
-            filters.append(User.receive_email_alerts == receive_email_alerts)
-        
-        # Query the database with the constructed filters
-        users = User.query.filter(*filters).all()
-        
-        # Check if no users were found
-        if not users:
-            return None
-
-        # Return list of email addresses
-        return [user.email for user in users]
 
 @app.route('/login', methods=['GET', 'POST'])
 def login():
@@ -144,105 +116,6 @@ def signup():
 
     return render_template('signup.html')
 
-@app.route('/protected')
-@login_required
-def protected():
-    if current_user.user_level == 'admin':
-        return f'Hello, Admin {current_user.username}! This is a protected page.'
-    return f'Hello, {current_user.username}! This is a protected page.'
-
-@app.route('/users')
-@login_required
-def view_users():
-    if current_user.user_level != 'admin':
-        flash("Your account does not have permission to view this page.", "danger")
-        return render_template("error/permission_denied.html")
-
-    # Fetch all users from the database
-    users = User.query.all()
-
-    return render_template('view_users.html', users=users)
-
-@app.route('/user/<username>', methods=['GET', 'POST'])
-@login_required
-def change_user_settings(username):
-    user = User.query.filter_by(username=username).first_or_404()
-
-    if request.method == 'POST':
-        new_username = request.form['username']
-        new_email = request.form['email']
-        new_user_level = request.form['user_level']
-        receive_email_alerts = 'receive_email_alerts' in request.form
-
-        # Update user details
-        user.username = new_username
-        user.email = new_email
-        user.user_level = new_user_level
-        user.receive_email_alerts = receive_email_alerts
-
-        db.session.commit()
-
-        flash('User settings updated successfully!', 'success')
-        return redirect(url_for('change_user_settings', username=user.username))
-
-    return render_template('change_user.html', user=user)
-
-@app.route('/delete_user/<username>', methods=['POST'])
-@login_required
-def delete_user(username):
-    if current_user.user_level != 'admin':
-        flash("Your account does not have permission to perform this action.", "danger")
-        return redirect(url_for('view_users'))  # Redirect to the users page
-
-    user = User.query.filter_by(username=username).first_or_404()
-
-    # Get Admin Emails with Alerts Enabled:
-    admin_email_address = get_email_addresses(user_level='admin', receive_email_alerts=True)
-    if admin_email_address:
-        subject = "User Deletion Alert"
-        context = {
-            "username": user.username,
-            "deletion_time": datetime.datetime.now(),
-            "current_user": current_user.username,
-        }
-        html_body = render_template_from_file("src/templates/email_templates/deletion_email.html", **context)
-        send_smpt_email(admin_email_address, subject, html_body, is_html=True)
-
-    db.session.delete(user)
-    db.session.commit()
-
-            
-    
-    flash(f'User {username} has been deleted successfully!', 'success')
-    return redirect(url_for('view_users'))
-
-@app.route("/update-email-password", methods=["GET", "POST"])
-@login_required
-def update_smpt_email_password():
-    smtp_config = SmptEamilPasswordConfig.query.first()
-
-    if request.method == "POST":
-        new_email = request.form.get("email")
-        new_password = request.form.get("password")
-
-        if not new_email or not new_password:
-            flash("Please provide email and password.", "danger")
-            return redirect(url_for("update_smpt_email_password"))
-        
-        
-        if not smtp_config:
-            smtp_config = SmptEamilPasswordConfig(email=new_email, password=new_password)
-            db.session.add(smtp_config)
-        else:
-            smtp_config.email = new_email
-            smtp_config.password = new_password        
-        
-        db.session.commit()
-        flash("Email and password updated successfully!", "success")
-        return redirect(url_for("update_smpt_email_password"))
-
-    return render_template("update_smpt_email_password.html", smtp_config=smtp_config)
-
 @app.route("/send_email", methods=["GET", "POST"])
 @login_required
 def send_email_page():
@@ -301,74 +174,3 @@ def send_email_page():
 
     return render_template("send_email.html", enable_alerts=enable_alerts)
 
-@app.route('/terminal', methods=['GET', 'POST'])
-@login_required
-def terminal():
-    if current_user.user_level != 'admin':
-        flash("Your account does not have permission to view this page.", "danger")
-        return render_template("error/permission_denied.html")
-    if request.method == 'POST':
-        command = request.form.get('command')
-        if command:
-            try:
-                # Run the command and capture the output
-                output = subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT, universal_newlines=True)
-            except subprocess.CalledProcessError as e:
-                # If the command fails, capture the error output
-                output = e.output
-            return jsonify(output=output)
-    return render_template('terminal.html')
-
-@app.route('/add_user', methods=['GET', 'POST'])
-@login_required
-def add_user():
-    if current_user.user_level != 'admin':
-        flash("Your account does not have permission to view this page.", "danger")
-        return render_template("error/permission_denied.html")
-    if request.method == 'POST':
-        username = request.form['username']
-        email = request.form['email']
-        password = request.form['password']
-        user_level = request.form.get('user_level', 'user')
-        receive_email_alerts = request.form.get('receive_email_alerts', 'on') == 'on'
-
-        # Check if user already exists
-        if User.query.filter_by(username=username).first() or User.query.filter_by(email=email).first():
-            flash('Username or email already exists.', 'danger')
-            return redirect(url_for('add_user'))
-
-        new_user = User(
-            username=username,
-            email=email,
-            password=generate_password_hash(password),
-            user_level=user_level,
-            receive_email_alerts=receive_email_alerts
-        )
-
-        admin_email_address = get_email_addresses(user_level='admin', receive_email_alerts=True)
-        if admin_email_address:
-            subject = "New User Alert"
-            context = {
-                "current_user": current_user.username,
-                "username": new_user.username,
-                "email": new_user.email,
-                "signup_time": datetime.datetime.now(),
-                "user_level": new_user.user_level
-            }
-            html_body = render_template_from_file("src/templates/email_templates/new_user_create.html", **context)
-            send_smpt_email(admin_email_address, subject, html_body, is_html=True)
-
-        subject = "Welcome to the systemGuard"  
-        context = {
-            "username": new_user.username,
-            "email": new_user.email,
-        }
-        html_body = render_template_from_file("src/templates/email_templates/welcome.html", **context)
-        send_smpt_email(email, subject, html_body, is_html=True)
-            
-        db.session.add(new_user)
-        db.session.commit()
-        flash('User created successfully!', 'success')
-        return redirect(url_for('view_users'))
-
-    return render_template('add_user.html')
