+++ src/static/css/dashboard_network.css
@@ -3,6 +3,7 @@
     margin: 0 auto;
     padding: 20px;
     text-align: center;
+    color: #fff;
 }
 
 .card-container {
@@ -13,7 +14,7 @@
 }
 
 .card {
-    background-color: #f8f9fa;
+    background-color: var(--color-username);
     border: 1px solid #ddd;
     border-radius: 8px;
     padding: 20px;
@@ -25,16 +26,25 @@
 }
 
 .card:hover {
-    transform: translateY(-5px);
+    transform: translateX(-10px); /* Slight lift effect */
+    box-shadow: 0 10px 20px #378fe7;
+    transition: transform 0.3s ease, box-shadow 0.3s ease; /* Smooth transition */
 }
 
 .card h2 {
     margin-top: 0;
     font-size: 1.5em;
+    color: #fff;
 }
 
 .card p {
     margin: 5px 0;
+    color: #fff;
+}
+
+.card a {
+    color: #fff;
+    text-decoration: none;
 }
 
 .add-server-link {
