+++ src/templates/add_server.html
@@ -47,7 +47,7 @@
                 <label for="type_selection">What do you have?</label>
                 <select id="type_selection" name="type_selection" onchange="toggleFields()">
                     <option value="ip">IP Address and Port</option>
-                    <option value="link">Link</option>
+                    <option value="link">URL</option>
                 </select>
             </div>
             <div id="ip_group" class="form-group">
@@ -59,7 +59,7 @@
                 <input type="number" id="port" name="port">
             </div>
             <div id="link_group" class="form-group">
-                <label for="link">Link</label>
+                <label for="link">URL</label>
                 <input type="text" id="link" name="link">
             </div>
             <button type="submit">Add Server</button>
